// src/generated/api/model/organisationRole.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * NestJS Boilerplate API
 * API documentation for the NestJS boilerplate project.
 * OpenAPI spec version: 1.0
 */

export type OrganisationRole = typeof OrganisationRole[keyof typeof OrganisationRole];


export const OrganisationRole = {
  OWNER: 'OWNER',
  MEMBER: 'MEMBER',
} as const;